+++ crates/api_server/src/handler/common.rs
@@ -64,7 +64,7 @@ pub async fn get_raw_object(
     app: &AppState,
     root_blob_name: &str,
     key: &str,
-    trace_id: Option<u64>,
+    trace_id: Option<u128>,
 ) -> Result<ObjectLayout, S3Error> {
     let nss_client = app.get_nss_rpc_client();
     let resp = nss_rpc_retry!(
@@ -102,7 +102,7 @@ pub async fn list_raw_objects(
     delimiter: &str,
     start_after: &str,
     skip_mpu_parts: bool,
-    trace_id: Option<u64>,
+    trace_id: Option<u128>,
 ) -> Result<Vec<(String, ObjectLayout)>, S3Error> {
     let nss_client = app.get_nss_rpc_client();
     let resp = nss_rpc_retry!(
